requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
deepgram-sdk==2.11.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
PyAudio==0.2.13
pycparser==2.21
requests==2.31.0
sounddevice==0.4.6
tqdm==4.66.1
urllib3==2.0.4
websockets==11.0.3
yarl==1.9.2
